revert(tui): unship `--simple` and delete the dead single-session REPL (#936)
Takes issue 936's option 1 after all: the three unreachable surfaces are
gone and the leaf helpers remain.

`--simple` shipped this morning as option 2 and reached no users. It was the
wrong shape for the problem it was sold as solving. Two things settled it:

- **It had nothing the deck lacks.** Every panel it drew, the deck draws —
  `render_hud`, `render_scope_review`, `render_ask_user`, and
  `render_slash_popup` are literally the same functions, called from
  `views/session.rs` and `deck_render.rs`; the composer, paste chips, and
  attachments are the same `crate::composer`. What was unique to it was a
  duplicate frame layout, a duplicate key handler, and a duplicate run loop —
  which is precisely what #936 was filed about.
- **A separate accessible surface is a permanent second-class tier.** Every
  deck feature after it would have been a new gap, against 25,886 lines of
  deck it would never catch. Accessibility belongs in the product.

The accessibility mechanisms were the valuable part and are not lost — they
are specced against the deck in the follow-up ticket, where the layout is
already single-column, the #935 cursor fix already ships, and `--no-anim`
already exists.

Deleted: `shell.rs`, `ui.rs`, the top-level `render` composer and the panels
only it drew (`render_transcript`, `render_files`, `render_diff`,
`render_composer`, `file_line`), `guarded_panel` and its helpers (the deck
uses `panel_guard`), `entry::transcript_lines` (its only caller was `ui.rs`;
the deck composes lanes itself), plus `simple_session`, the demo example, and
the pty test.

Kept, because the deck renders through them: `entry_lines`, `render_hud`,
`render_transcript_window`, `render_scope_review`, `render_ask_user`,
`render_slash_popup`, `slash_popup_area`, `scroll_window_start`, `justify`,
`human_duration`, and the inline-diff resolvers.

`DebugLog` and the secure `append_json_line` moved to their own module: they
were declared in `shell.rs` but never belonged to it — the deck writes through
them and the panic hook appends to the same file.

Test coverage was audited rather than deleted wholesale. Of the surface's
tests, the 22 that assert on `entry_lines` output (inline diffs, the brand
palette, slash-popup windowing, collapsed reasoning) were kept and re-parented;
the ones that built a `UiState` and drew a frame went with the frame. The
transcript-composition helper they share is now a test fixture, so the
assertions survive without a production function nothing calls.

fmt, clippy -D warnings, rustdoc -D warnings, and the wire gate all pass; 698
stella-tui and 1133 stella-cli tests green.

Net: 370 insertions, 5,268 deletions.
