+++ stella-tui/src/render/tests/slash.rs
@@ -10,76 +10,6 @@
 use super::*;
 
 #[test]
-fn slash_menu_renders_filtered_commands() {
-    let model = SessionModel::new();
-    let mut composer = Composer::new();
-    for c in "/di".chars() {
-        composer.insert_char(c);
-    }
-    let mut ui = UiState::new(
-        composer,
-        vec![
-            SlashCommand::new("/diff", "open the diff viewer"),
-            SlashCommand::new("/files", "focus files"),
-        ],
-    );
-    let text = draw(&model, &mut ui, 100, 30);
-    assert!(text.contains("/diff"), "slash menu lists /diff:\n{text}");
-    assert!(!text.contains("/files"), "filtered out /files:\n{text}");
-}
-
-#[test]
-fn slash_popup_marks_the_selected_command() {
-    let model = SessionModel::new();
-    let mut composer = Composer::new();
-    composer.insert_char('/');
-    let mut ui = UiState::new(
-        composer,
-        vec![
-            SlashCommand::new("/help", "show help"),
-            SlashCommand::new("/diff", "open the diff viewer"),
-        ],
-    );
-    ui.slash_selected = 1;
-    let text = draw(&model, &mut ui, 100, 30);
-    // The glyph is double-width: its trailing filler cell dumps as an
-    // extra space before the explicit separator.
-    assert!(
-        text.contains("▸ 🔒  /diff"),
-        "selection marker + builtin glyph on the second row:\n{text}"
-    );
-    assert!(
-        text.contains("/ commands · 2"),
-        "popup title with the match count:\n{text}"
-    );
-}
-
-#[test]
-fn slash_popup_glyphs_distinguish_builtin_from_custom_commands() {
-    let model = SessionModel::new();
-    let mut composer = Composer::new();
-    composer.insert_char('/');
-    let mut ui = UiState::new(
-        composer,
-        vec![
-            SlashCommand::new("/help", "show help"),
-            SlashCommand::custom("/fix-bug", "fix the named bug"),
-        ],
-    );
-    let text = draw(&model, &mut ui, 100, 30);
-    assert!(
-        text.contains("🔒  /help"),
-        "productized commands carry the lock glyph:\n{text}"
-    );
-    assert!(
-        text.contains("⚡  /fix-bug"),
-        "custom commands carry the lightning glyph:\n{text}"
-    );
-}
-
-// ---- Slash-popup windowing ---------------------------------------------
-
-#[test]
 fn scroll_window_start_holds_still_until_the_selection_leaves_the_edge() {
     // Fits entirely: never scrolls.
     assert_eq!(scroll_window_start(5, 4, 8), 0);
