+++ website/content/docs/commands/chat.mdx
@@ -1,11 +1,11 @@
 ---
 title: stella chat
-description: Start an interactive session with Stella — the Command Deck TUI by default, an accessible single-pane surface, or a plain line REPL.
+description: Start an interactive session with Stella — the Command Deck TUI by default, or a plain line REPL.
 ---
 
 Start an interactive session. This is the **default** command: running `stella` with no subcommand is exactly the same as running `stella chat`.
 
-On a real terminal, `stella chat` opens the **Command Deck** — a tabbed terminal UI. There are two other surfaces: `--simple` (or `STELLA_SIMPLE=1`) for the single-pane surface built for screen readers, and `--plain` (or `STELLA_PLAIN=1`) for a line-based REPL. The deck also steps aside automatically when stdin or stdout is not a terminal.
+On a real terminal, `stella chat` opens the **Command Deck** — a tabbed terminal UI. Pass `--plain` (or set `STELLA_PLAIN=1`) for a plain line-based REPL instead; the deck also steps aside automatically when stdin or stdout is not a terminal.
 
 ## Synopsis
 
@@ -15,9 +15,6 @@ stella chat [global flags]
 # Equivalent — chat is the default
 stella [global flags]
 
-# Single-pane accessible surface (screen readers, limited terminals)
-stella chat --simple
-
 # Plain line-based REPL instead of the deck
 stella chat --plain
 ```
@@ -26,7 +23,7 @@ stella chat --plain
 
 `stella chat` opens a conversational session. You type a message, the agent runs its step loop — proposing tool calls, executing them, and feeding results back until the turn is done — and then hands the prompt back to you. Conversation history is retained across turns within the session.
 
-The default **Command Deck** presents the session as a set of tabs — **SESSION** (the transcript), **AGENTS** (executions & installed agents), **TRACES**, **GRAPH**, **FILES**, **SKILLS**, **MCP**, **ISSUES** (your connected GitHub/Linear tracker, with instant type-ahead), and **SETTINGS** (the home of all config, incl. the engine-config panel). Slash commands jump between tabs and run actions, and `?` opens a tab-aware key list. The **simple surface** (`--simple`) is one conversation drawn inline, built for screen readers. The **plain REPL** (`--plain`) is a single scrolling line-based loop with a smaller command set — handy for pipes, recordings, and terminals that cannot do better.
+The default **Command Deck** presents the session as a set of tabs — **SESSION** (the transcript), **AGENTS** (executions & installed agents), **TRACES**, **GRAPH**, **FILES**, **SKILLS**, **MCP**, **ISSUES** (your connected GitHub/Linear tracker, with instant type-ahead), and **SETTINGS** (the home of all config, incl. the engine-config panel). Slash commands jump between tabs and run actions, and `?` opens a tab-aware key list. The **plain REPL** (`--plain`) is a single scrolling line-based loop with a smaller command set — handy for narrow terminals, screen readers, and recordings.
 
 ### Typing while a turn is running
 
@@ -79,15 +76,9 @@ All [global flags](/docs/commands) apply, in either position — `stella --model
 `stella chat` takes only the [global flags](/docs/commands). The ones that shape the chat surface:
 
 <CardGrid size="md">
-  <OptionCard name="--simple">
-    Use the single-pane accessible surface instead of the Command Deck. Env
-    `STELLA_SIMPLE=1`. Needs a terminal on stdin and stdout; over a pipe it falls back
-    to `--plain` and says which stream was the reason.
-  </OptionCard>
   <OptionCard name="--plain">
     Use the plain line-based REPL instead of the Command Deck. Env `STELLA_PLAIN=1`; also
-    auto-selected on a non-TTY. A typed `--simple` wins over an inherited `STELLA_PLAIN`;
-    a typed `--plain` wins over everything.
+    auto-selected on a non-TTY.
   </OptionCard>
   <OptionCard name="--no-anim">
     Freeze deck animation (progress shimmer, caret blink) to a static frame — for
@@ -234,43 +225,6 @@ Clamping alone would flatten the ladder — on a two-rung provider both `fast` a
 
 Like `/model`, the change applies to sessions started from now on; the running session keeps its current model.
 
-### Simple surface (`--simple`)
-
-Built for screen readers, and useful to anyone on a terminal that cannot do
-better. It is one conversation, one turn at a time — no tabs, no prompt queue,
-no sub-agents, no resume. Run `stella chat` when you want those.
-
-What makes it different from the deck is not that it has fewer features:
-
-- **It never takes over the screen.** The deck draws on the *alternate screen*,
-  a separate grid it rewrites wholesale several times a second. A screen reader
-  has no way to tell that from "something changed somewhere", so it announces
-  either nothing or everything. The simple surface draws inline, underneath
-  whatever was already in your terminal.
-- **Finished messages become ordinary terminal output.** As each message
-  settles it moves into your scrollback — written once, announced once as it
-  arrives, and reachable afterwards with your reader's review cursor or your
-  terminal's own scroll. Quit and it is all still there.
-- **Reading order is document order.** The panels stack in one column. A screen
-  reader walks the grid row by row, so the deck's side-by-side layout gives you
-  half a sentence of transcript followed by half a filename on every row.
-- **The terminal cursor sits on the real insertion point.** That is what a
-  reader follows, and what a CJK/Japanese/Korean input method anchors its
-  candidate window to — without it, composition appears in the wrong place or
-  not at all.
-- **The mouse is never captured**, so your terminal's own selection keeps
-  working.
-
-Keys are the deck's: **Enter** sends, **Esc** interrupts a running turn,
-**Ctrl-C** quits, **Tab** moves between the transcript and the files list,
-**Ctrl-R** expands reasoning, and **/** opens the command menu. Its command set
-is `/help`, `/clear`, `/files`, and `/exit`, plus any custom commands you have
-defined.
-
-Set `STELLA_SIMPLE=1` to make it the default without typing the flag. It needs
-a real terminal on both stdin and stdout; over a pipe it falls back to the plain
-REPL and names the stream that caused it.
-
 ### Plain REPL (`--plain`)
 
 The line REPL has a smaller, different set:
@@ -331,19 +285,7 @@ Pin a model and cap session spend:
 stella --model zai/glm-5.2 --budget 5.00 chat
 ```
 
-Use the accessible single-pane surface (screen readers, limited terminals):
-
-```bash
-stella chat --simple
-```
-
-Make it the default for every session, without typing the flag:
-
-```bash
-export STELLA_SIMPLE=1
-```
-
-Use the plain line REPL (a pipe, a CI log, a terminal that cannot do better):
+Use the plain line REPL (e.g. for a screen reader or a narrow terminal):
 
 ```bash
 stella chat --plain
